--- cg/meta/orders/__init__.py
@@ -1,1 +1,0 @@
-from .api import OrdersAPI
